+++ assets/js/components/catalog.vue
@@ -55,10 +55,17 @@ export default {
             legend: 'Shipping takes 10-12 weeks, and products probably won\'t work',
             products: [], // Reactive array for products
             loading: false, // tracks loading status // added flag for tracking Ajax call
+            searchTerm: null, // state for search term, now we can watch this property for changes and react to it by reloading products with the new search term.
         };
     },
+    watch: { // watchers allow us to react to changes in reactive data properties. In this case, we want to watch for changes in the currentCategoryId, so that when the user clicks on a different category in the sidebar, we can react to that change and load the products for the new category.
+        currentCategoryId() {
+            // Reload products when category changes
+            this.loadProducts();
+        },
+    },
     mounted() {
-        this.loadProducts(null); // Load products when component is mounted, with no search term (null)
+        this.loadProducts(); // Load products when component is mounted, with no search term
     },
     methods: {
         // onSearchProducts(event) { // look at the shorter example below, we can use destructuring to extract term directly from the event object, which makes the code cleaner and more concise.
@@ -69,16 +76,17 @@ export default {
        * @param { string } term
        */
         onSearchProducts({ term }) {
+            this.searchTerm = term;
             // Update local searchTerm based on emitted payload
-            this.loadProducts(term); // Load products with the new search term
+            this.loadProducts(); // Load products with the new search term
         },
-        async loadProducts(searchTerm) {
+        async loadProducts() {
             this.loading = true; // Start loading state
 
             let response = null;
             // If a category ID is provided, fetch products for that category
             try {
-                response = await fetchProducts(this.currentCategoryId, searchTerm); // Use the service function to fetch products with optional category filter
+                response = await fetchProducts(this.currentCategoryId, this.searchTerm); // Use the service function to fetch products with optional category filter
             // .log(response); // Log full Axios response (headers, status, data, etc.)
             } catch (error) {
                 console.error('Error fetching products:', error);